pyproject.toml
[project]
name = "image_editor"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ethanposey", email = "ethanposeyk@outlook.com"},
]
dependencies = ["pillow>=11.2.1"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}


[dependency-groups]
lint = [
    "ruff>=0.11.11",
    "mypy>=1.15.0",
]
pytest = [
    "pytest-cov>=6.1.1",
]

[tool.ruff]
select = [
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "E", # pycodestyle - Error
    "F", # Pyflakes
    "I", # isort
    "W", # pycodestyle - Warning
    "UP", # pyupgrade
]
ignore = [
    "E501", # line-too-long
    "W191", # tab-indentation
]
include = ["**/*.py", "**/*.pyi", "**/pyproject.toml"]
[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
strict = true

[tool.pdm]
distribution = false

[tool.pdm.scripts]
lint = "ruff ."
fmt = "ruff format ."